+++ src/core/QueryManager.ts
@@ -1333,7 +1333,6 @@ export class QueryManager<TStore> {
       returnPartialData,
       context,
       notifyOnNetworkStatusChange,
-      fetchBlockingPromise,
     }: WatchQueryOptions<TVars, TData>,
     // The initial networkStatus for this fetch, most often
     // NetworkStatus.loading, but also possibly fetchMore, poll, refetch,
@@ -1391,41 +1390,16 @@ export class QueryManager<TStore> {
       ) ? CacheWriteBehavior.OVERWRITE
         : CacheWriteBehavior.MERGE;
 
-    const resultsFromLink = () => {
-      const get = () => this.getResultsFromLink<TData, TVars>(
-        queryInfo,
-        cacheWriteBehavior,
-        {
-          variables,
-          context,
-          fetchPolicy,
-          errorPolicy,
-        },
-      );
-
-      // If we have a fetchBlockingPromise, wait for it to be resolved before
-      // allowing any network requests, and only proceed if fetchBlockingPromise
-      // resolves to true. If it resolves to false, the request is discarded.
-      return fetchBlockingPromise ? fetchBlockingPromise.then(
-        ok => ok ? get() : Observable.of<ApolloQueryResult<TData>>(),
-        error => {
-          const apolloError = isApolloError(error)
-            ? error
-            : new ApolloError({ clientErrors: [error] });
-
-          if (errorPolicy !== "ignore") {
-            queryInfo.markError(apolloError);
-          }
-
-          return Observable.of<ApolloQueryResult<TData>>({
-            loading: false,
-            networkStatus: NetworkStatus.error,
-            error: apolloError,
-            data: readCache().result,
-          });
-        },
-      ) : get();
-    }
+    const resultsFromLink = () => this.getResultsFromLink<TData, TVars>(
+      queryInfo,
+      cacheWriteBehavior,
+      {
+        variables,
+        context,
+        fetchPolicy,
+        errorPolicy,
+      },
+    );
 
     const shouldNotify =
       notifyOnNetworkStatusChange &&