+++ src/react/hooks/useQuery.ts
@@ -1,7 +1,6 @@
 import {
   useCallback,
   useContext,
-  useEffect,
   useMemo,
   useRef,
   useState,
@@ -212,36 +211,22 @@ class InternalState<TData, TVariables> {
     // allows us to depend on the referential stability of
     // this.watchQueryOptions elsewhere.
     const currentWatchQueryOptions = this.watchQueryOptions;
-    let resolveFetchBlockingPromise: undefined | ((result: boolean) => any);
-
     if (!equal(watchQueryOptions, currentWatchQueryOptions)) {
       this.watchQueryOptions = watchQueryOptions;
       if (currentWatchQueryOptions && this.observable) {
-        // Though it might be tempting to postpone this setOptions call to the
+        // Though it might be tempting to postpone this reobserve call to the
         // useEffect block, we need getCurrentResult to return an appropriate
         // loading:true result synchronously (later within the same call to
         // useQuery). Since we already have this.observable here (not true for
         // the very first call to useQuery), we are not initiating any new
-        // subscriptions, though it does feel less than ideal that setOptions
+        // subscriptions, though it does feel less than ideal that reobserve
         // (potentially) kicks off a network request (for example, when the
-        // variables have changed). To prevent any risk of premature/unwanted
-        // network traffic, we use a fetchBlockingPromise, which will only be
-        // unblocked once the useEffect has fired.
-        this.observable.reobserve({
-          fetchBlockingPromise: new Promise<boolean>(resolve => {
-            resolveFetchBlockingPromise = resolve;
-          }),
-          // If watchQueryOptions.fetchBlockingPromise is also defined, it takes
-          // precedence over the fetchBlockingPromise we just created.
-          ...watchQueryOptions,
-        });
-
+        // variables have changed), which is technically a side-effect.
+        this.observable.reobserve(watchQueryOptions);
         this.previousData = this.result?.data || this.previousData;
         this.result = void 0;
       }
     }
-
-    useUnblockFetchEffect(this.renderPromises, resolveFetchBlockingPromise);
 
     // Make sure state.onCompleted and state.onError always reflect the latest
     // options.onCompleted and options.onError callbacks provided to useQuery,
@@ -424,8 +409,6 @@ class InternalState<TData, TVariables> {
   >;
 
   private useObservableQuery() {
-    let resolveFetchBlockingPromise: undefined | ((result: boolean) => any);
-
     // See if there is an existing observable that was used to fetch the same
     // data and if so, use it instead since it will contain the proper queryId
     // to fetch the result set. This is used during SSR.
@@ -434,13 +417,8 @@ class InternalState<TData, TVariables> {
         && this.renderPromises.getSSRObservable(this.watchQueryOptions)
         || this.observable // Reuse this.observable if possible (and not SSR)
         || this.client.watchQuery({
-          fetchBlockingPromise: new Promise<boolean>(resolve => {
-            resolveFetchBlockingPromise = resolve;
-          }),
           ...this.watchQueryOptions,
         });
-
-    useUnblockFetchEffect(this.renderPromises, resolveFetchBlockingPromise);
 
     this.obsQueryFields = useMemo(() => ({
       refetch: obsQuery.refetch.bind(obsQuery),
@@ -565,27 +543,4 @@ class InternalState<TData, TVariables> {
       this.observable.refetch();
     }
   }
-}
-
-function useUnblockFetchEffect<TData, TVars>(
-  renderPromises: ApolloContextValue["renderPromises"],
-  resolveFetchBlockingPromise?: (result: boolean) => any,
-) {
-  if (resolveFetchBlockingPromise) {
-    if (renderPromises) {
-      // Since we're doing SSR, the useEffect callback will not be called, so we
-      // must unblock the fetchBlockingPromise now.
-      resolveFetchBlockingPromise(true);
-    } else {
-      // Otherwise, silently discard blocked fetches whose useEffect callbacks
-      // have not fired within 5 seconds (more than enough time to mount).
-      setTimeout(() => resolveFetchBlockingPromise(false), 5000);
-    }
-  }
-
-  useEffect(() => {
-    if (resolveFetchBlockingPromise) {
-      resolveFetchBlockingPromise(true);
-    }
-  }, [resolveFetchBlockingPromise]);
 }